cvs merge + GargoyleMT's patch + other fixes

+++ client/BufferedSocket.h
@@ -135,7 +135,7 @@ public:
 	}
 	
 	void write(const string& aData) throw(SocketException) { write(aData.data(), aData.length()); };
-	virtual void write(const char* aBuf, int aLen) throw();
+	virtual void write(const char* aBuf, size_t aLen) throw();
 
 	/**
 	 * Send the file f over this socket.
@@ -175,8 +175,8 @@ private:
 	size_t inbufSize;
 	enum {BUFFERS = 2};
 	u_int8_t* outbuf[BUFFERS];
-	int outbufSize[BUFFERS];
-	int outbufPos[BUFFERS];
+	size_t outbufSize[BUFFERS];
+	size_t outbufPos[BUFFERS];
 	int curBuf;
 
 	InputStream* file;
