+++ client/SearchManager.cpp
@@ -164,7 +164,7 @@ int SearchManager::run() {
 	return 0;
 }
 
-void SearchManager::onData(const u_int8_t* buf, int aLen, const string& address) {
+void SearchManager::onData(const u_int8_t* buf, size_t aLen, const string& address) {
 	string x((char*)buf, aLen);
 	if(x.compare(0, 4, "$SR ") == 0) {
 		onNMDCData(buf, aLen, address);
@@ -218,7 +218,7 @@ string SearchManager::clean(const string& aSearchString) {
 	return tmp;
 }
 
-void SearchManager::onNMDCData(const u_int8_t* buf, int aLen, const string& address) {
+void SearchManager::onNMDCData(const u_int8_t* buf, size_t aLen, const string& address) {
 	string x((char*)buf, aLen);
 		string::size_type i, j;
 		// Directories: $SR <nick><0x20><directory><0x20><free slots>/<total slots><0x05><Hubname><0x20>(<Hubip:port>)
