+++ windows/DirectoryListingFrm.h
@@ -306,7 +306,7 @@ private:
 
 	int skipHits;
 
-	int files;
+	size_t files;
 
 	bool updating;
 	bool searching;
